Grand Finale
Aside from (highly probable) bug fixes, this is most likely the finished
version of the program.  It does not have any graphics- which is
unfortunate, seeing as they're essential to understanding the game- but
allows for the complete range of actions that would go into playing.
You can even win!... if you can figure out what's going on from the text
interface.